refactored error
